# swagger.yaml
swagger: "2.0"
info:
  description: ""
  version: "1.0.0"
  title: "API IoT"
...
schemes:
- "https"
paths:
...
  /user:
    post:
      tags:
      - "user"
      summary: "Create user"
      description: "Permet de créer un nouvel utilisateur dans la base de données"
      operationId: "createUser"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "User au format JSON"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        200:
          description: "successful operation"
  /user/login:
    get:
      tags:
      - "user"
      summary: "Connecte un utilisateur"
      description: ""
      operationId: "loginUser"
      parameters:
      - name: "username"
        in: "query"
        description: "Nom d'utilisateur"
        required: true
        type: "string"
      - name: "password"
        in: "query"
        description: "Mot de passe"
        required: true
        type: "string"
      responses:
        "200":
          description: "successful operation"
        "400":
          description: "Invalid username/password supplied"
...
  /user/{username}:
    get:
      tags:
      - "user"
      summary: "Récupère un utilisateur par son nom d'utilisateur"
      description: ""
      operationId: "getUserByName"
      produces:
      - "application/json"
      parameters:
      - name: "username"
        in: "path"
        description: "Le nom de l'utilisateur à récupérer"
        required: true
        type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/UserPrivate"
        "404":
          description: "User not found"
    delete:
      tags:
      - "user"
      summary: "Supprime un utilisateur"
      description: "Appel uniquement possible si l'utilisateur est connecté."
      operationId: "deleteUser"
      produces:
      - "application/json"
      parameters:
      - name: "username"
        in: "path"
        description: "The name that needs to be deleted"
        required: true
        type: "string"
      responses:
        "400":
          description: "Invalid username supplied"
        "404":
          description: "User not found"
...
securityDefinitions:
  petstore_auth:
    type: "oauth2"
    authorizationUrl: "http://petstore.swagger.io/oauth/dialog"
    flow: "implicit"
    scopes:
      write:pets: "modify pets in your account"
      read:pets: "read your pets"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
...
  User:
    type: "object"
    properties:
      username:
        type: "string"
      password:
        type: "string"
    xml:
      name: "User"
  UserPrivate:
    type: "object"
    properties:
      username:
        type: "string"
    xml:
      name: "User"
...